+++ fe/fe-core/src/main/java/com/starrocks/connector/delta/DeltaUtils.java
@@ -82,7 +82,7 @@ public class DeltaUtils {
             fullSchema.add(column);
         }
 
-        return new DeltaLakeTable(CONNECTOR_ID_GENERATOR.getNextId().asInt(), catalog, dbName, tblName, fullSchema,
+        return new DeltaLakeTable(CONNECTOR_ID_GENERATOR.getNextId().asLong(), catalog, dbName, tblName, fullSchema,
                 loadPartitionColumnNames(snapshotImpl), snapshotImpl, deltaLakeEngine, snapshot.getMetastoreTable());
     }
