+++ fe/fe-core/src/main/java/com/starrocks/connector/metadata/iceberg/IcebergManifestsTable.java
@@ -44,7 +44,7 @@ public class IcebergManifestsTable extends MetadataTable {
 
     public static IcebergManifestsTable create(String catalogName, String originDb, String originTable) {
         return new IcebergManifestsTable(catalogName,
-                ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asInt(),
+                ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asLong(),
                 TABLE_NAME,
                 Table.TableType.METADATA,
                 builder()
