+++ fe/fe-core/src/main/java/com/starrocks/connector/metadata/iceberg/LogicalIcebergMetadataTable.java
@@ -44,7 +44,7 @@ public class LogicalIcebergMetadataTable extends MetadataTable {
 
     public static LogicalIcebergMetadataTable create(String catalogName, String originDb, String originTable) {
         return new LogicalIcebergMetadataTable(catalogName,
-                ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asInt(),
+                ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asLong(),
                 TABLE_NAME,
                 Table.TableType.METADATA,
                 builder()
