+++ fe/fe-core/src/main/java/com/starrocks/server/CatalogMgr.java
@@ -120,7 +120,7 @@ public class CatalogMgr {
                     throw new DdlException("connector create failed");
                 }
                 long id = isResourceMappingCatalog(catalogName) ?
-                        ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asInt() :
+                        ConnectorTableId.CONNECTOR_ID_GENERATOR.getNextId().asLong() :
                         GlobalStateMgr.getCurrentState().getNextId();
                 catalog = new ExternalCatalog(id, catalogName, comment, properties);
                 catalogs.put(catalogName, catalog);
